Added support for automatically making navbar elements "active" and removed ID from Users table, Username is the primary key now.
The navbar thing works by looking up the associated route's name with the name of the Handler, then it uses some jquery to add the CSS class "active" to it by getting the current page's handler via some magical built in element. It is kind of of bad but it works!

This hasn't happened to the admin page yet, I should really properly template the pills on that.