Started work on a dashboard. Only shows process stats so far
The CPU progress bar is system-wide because of some strange issues with psutil and getting the CPU usage of threads.

It'd return 0 a lot of the time then randomly spit out an impossibly high number. Getting system CPU usage seems to be a lot more reliable.

PMCP-7 #Comment Started work on the dashboard, shows process statistics right now.